+++ app/src/main/java/org/ipforsmartobjects/apps/popularmovies/data/MoviesServiceApiImplWithoutRx.java
@@ -79,7 +79,7 @@ public class MoviesServiceApiImplWithoutRx implements MoviesServiceApi {
     }
 
     @Override
-    public void getMovie(int movieId, final MoviesServiceCallback<Movie> callback) {
+    public void getMovie(long movieId, final MoviesServiceCallback<Movie> callback) {
         // TODO: 3/5/2017 implement movie detail
         Call<Movie> movie = mApi.getMovieWithId(movieId, mApiKey);
