+++ app/src/main/java/org/ipforsmartobjects/apps/popularmovies/detail/MovieDetailContract.java
@@ -23,6 +23,6 @@ public interface MovieDetailContract {
     }
 
     interface UserActionsListener {
-        void openMovie(@Nullable int movieId);
+        void openMovie(@Nullable long movieId);
     }
 }
